Add util function to parse arpa name to IpAddr
This commit adds a new util function to parse back "4.3.2.1.in-addr.arpa" to the
corresponding std::net::IpAddr.

This is usually needed when parsing the answer of reverse DNS queries. When
sending multiple rDNS queries in a short amount of time, because of UDP,
responses may arrive in the wrong order. In this case we need to parse the ip
address in the name field of the answer to find out which question is being answered.

I didn't really know what to return when the name does not describe a valid ip
address, so I introduced a new ParserError variant. ParserError::AddrError would
have been more appropriate, but it wraps a std::net::AddrParseError,
and the constructor of this type is private.