Moved display implementations under a new feature
display.rs implements the Display trait on a few structures in a dig style,
and pulls the chrono dependency for it, which is a pretty big crate.

These Display implementations are only needed for the dig client, and could
be skipped when the user is only interested in raw serialization and
deserialization of DNS messages.

This commit puts the whole display.rs module under a new feature flag, to
prevent pulling unnecessary crates when they are not needed.